pyproject.toml
[project]
name = "backend-carnet-digital"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=46.0.3",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-marshmallow>=1.3.0",
    "flask-migrate>=4.1.0",
    "flask-restful>=0.3.10",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "marshmallow-sqlalchemy>=1.4.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
]
